refactor(edge/stacks): move shared logic
edge-stacks shared logic - stack folder
